version finale

+++ libft/libft.h
@@ -34,7 +34,7 @@ char	*ft_strchr(const char *string, int c);
 char	*ft_strdup(const char *s);
 size_t	ft_strlcat(char *dest, const char *src, size_t dstsize);
 size_t	ft_strlcpy(char *dest, const char *src, size_t dstsize);
-int	    ft_strlen(const char *s);
+size_t	    ft_strlen(const char *s);
 int		ft_strncmp(const char *s1, const char *s2, size_t size);
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len);
 char	*ft_strrchr(const char *string, int c);
